Add roster endpoint, player list now connects to API, implemented selection and sorting logic

+++ FrozenFantasy/Components/FantasyPointLabel.swift
@@ -8,9 +8,9 @@
 import SwiftUI
 
 struct FantasyPointLabel: View {
-    let value: Int
+    let value: Double
 
-    init(_ value: Int) {
+    init(_ value: Double) {
         self.value = value
     }
